+++ include/SuperSoco485.h
@@ -15,8 +15,11 @@ namespace stprograms::SuperSoco485
      * @{
      */
 
+    // forward declaration
+    class SuperSoco485;
+
     /// @brief Callback definition on parsed telegram
-    typedef void (*ParsedTelegramReceived)(void* user_data, void *sender, BaseTelegram *telegram);
+    typedef void (*DataChangedHandler)(void *user_data, SuperSoco485 *sender);
 
     /// @brief Basic class for SuperSoco485
     class SuperSoco485
@@ -24,17 +27,42 @@ namespace stprograms::SuperSoco485
     public:
         SuperSoco485();
 
-        void setCallback(ParsedTelegramReceived callback, void *user_data);
+        void setCallback(DataChangedHandler callback, void *user_data);
         void begin();
         void update();
 
-    private:
-        ParsedTelegramReceived _callback;
-        uint8_t _rawBuffer[256];
-        stprograms::SuperSoco485::TelegramParser _parser;
-        void *user_data;
+        /// @brief Status information
+        class Status
+        {
+        public:
+            /// @brief Current battery voltage
+            byte batVoltage;
+            /// @brief State of Charge
+            byte Soc;
+            /// @brief battery temperature
+            byte batTemp;
+            /// @brief charge / discharge current
+            double chargeCurrent;
+            /// @brief Number of charging cycles
+            uint16_t chargeCycle;
+            /// @brief Currently charging
+            bool charging;
+        };
 
-        static void telegramParsed(void *sender, BaseTelegram *data);
+        /// @brief Status structure
+        Status _status;
+
+    private:
+        /// @brief function to call if data has changed
+        DataChangedHandler _callback;
+        /// @brief Userdata do transmit with callback function
+        void *_user_data;
+        /// @brief buffer of raw data for RS485
+        uint8_t _rawBuffer[256];
+        /// @brief telegram parser instance
+        stprograms::SuperSoco485::TelegramParser _parser;
+
+        static void batteryStatusReceived(void *sender, BaseTelegram *data);
     };
 }