+++ src/SuperSoco485.cpp
@@ -2,6 +2,8 @@
 #include <Arduino.h>
 #include <ArduinoRS485.h>
 
+#include "BatteryStatus.h"
+
 using namespace stprograms::SuperSoco485;
 
 /**
@@ -11,26 +13,27 @@ using namespace stprograms::SuperSoco485;
 
 /// @brief Baudrate used for communication
 const unsigned long SUPER_SOCO_BAUDRATE = 9600;
-const uint8_t TELEGRAM_TERMINATOR = 0x0D;
 
 /**
  * @brief Create a new instance of the SuperSoco485 class
- * @param callback Callback to register for parsed telegrams
  */
-SuperSoco485::SuperSoco485() 
-: _parser (stprograms::SuperSoco485::TelegramParser(telegramParsed, this))
+SuperSoco485::SuperSoco485()
+    : _parser(stprograms::SuperSoco485::TelegramParser(this))
 {
-
+    _parser.setBatStatusHandler(batteryStatusReceived);
 }
 
-void SuperSoco485::setCallback(ParsedTelegramReceived callback, void *user_data)
+/// @brief Set the data changed callback
+/// @param callback callback to be called
+/// @param user_data user data that will be transmitted with the callback
+void SuperSoco485::setCallback(DataChangedHandler callback, void *user_data)
 {
     if (callback != NULL)
     {
         this->_callback = callback;
     }
-    
-    this->user_data = user_data;
+
+    this->_user_data = user_data;
 }
 
 /**
@@ -56,7 +59,7 @@ void SuperSoco485::update()
         size_t readBytes = 0;
         do
         {
-            readBytes = RS485.readBytesUntil(TELEGRAM_TERMINATOR,
+            readBytes = RS485.readBytesUntil(TelegramParser::TELEGRAM_TERMINATOR,
                                              _rawBuffer,
                                              sizeof(_rawBuffer));
 
@@ -66,18 +69,46 @@ void SuperSoco485::update()
     }
 }
 
-void SuperSoco485::telegramParsed(void *sender, BaseTelegram *data)
+/// @brief Template function for comparing data. Sets the new value in current
+/// value and if the values had a different value, sets the hasChanged value to true
+/// @tparam T type of values to compare
+/// @param curVal current value to compare
+/// @param newVal new value to compare
+/// @param hasChanged value set to true if values have changed
+template <typename T>
+void compareData(T &curVal, T newVal, bool &hasChanged)
 {
-    SuperSoco485 *ss = (SuperSoco485*) sender;
-    
-    // TODO: Check for changes
+    if (curVal != newVal)
+    {
+        curVal = newVal;
+        hasChanged = true;
+    }
+}
+
+/// @brief A telegram has been parsed
+/// @param user_data pointer to SuperSoco485 instance
+/// @param data parsed telegram
+void SuperSoco485::batteryStatusReceived(void *user_data, BaseTelegram *data)
+{
+    SuperSoco485 *ss = (SuperSoco485 *)user_data;
+    BatteryStatus *status = (BatteryStatus *)data;
+    bool hasChanged = false;
+
+    // compare data
+    Serial.println(data->toString());
+
+    compareData(ss->_status.batVoltage, status->getVoltage(), hasChanged);
+    compareData(ss->_status.Soc, status->getSoC(), hasChanged);
+    compareData(ss->_status.batTemp, status->getTemperature(), hasChanged);
+    compareData(ss->_status.chargeCurrent, status->getChargeCurrent(), hasChanged);
+    compareData(ss->_status.chargeCycle, status->getCycles(), hasChanged);
+    compareData(ss->_status.charging, status->isCharging(), hasChanged);
 
     // Call callback
-    if(ss->_callback != NULL && data != NULL)
+    if (hasChanged && ss->_callback != NULL)
     {
-        ss->_callback(ss->user_data, ss, data);
+        ss->_callback(ss->_user_data, ss);
     }
-
 }
 
 /** @} */