+++ CHANGELOG.md
@@ -10,3 +10,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [0.1.0] - 2023-03-05
 
 ### Added
+- Implemented TelegramParser
+- Implemented BatteryStatus
+- Implemented SuperSoco485 base class